+++ album/get_tracks.go
@@ -3,9 +3,9 @@ package album
 import (
 	"encoding/json"
 	"errors"
-	"github.com/najemi-software/tidal-dl/v2/generics"
-	"github.com/najemi-software/tidal-dl/v2/requests"
-	"github.com/najemi-software/tidal-dl/v2/track"
+	"github.com/najemi-software/tidal-dl/v3/generics"
+	"github.com/najemi-software/tidal-dl/v3/requests"
+	"github.com/najemi-software/tidal-dl/v3/track"
 	"strconv"
 )