+++ artist/get_similar_artists.go
@@ -3,8 +3,8 @@ package artist
 import (
 	"encoding/json"
 	"errors"
-	"github.com/najemi-software/tidal-dl/v2/generics"
-	"github.com/najemi-software/tidal-dl/v2/requests"
+	"github.com/najemi-software/tidal-dl/v3/generics"
+	"github.com/najemi-software/tidal-dl/v3/requests"
 	"strconv"
 )