+++ search/get.go
@@ -3,7 +3,7 @@ package search
 import (
 	"encoding/json"
 	"errors"
-	"github.com/najemi-software/tidal-dl/v2/requests"
+	"github.com/najemi-software/tidal-dl/v3/requests"
 	"net/url"
 	"strconv"
 	"strings"