+++ session/get.go
@@ -3,7 +3,7 @@ package session
 import (
 	"encoding/json"
 	"errors"
-	"github.com/najemi-software/tidal-dl/v2/requests"
+	"github.com/najemi-software/tidal-dl/v3/requests"
 	"strconv"
 )