+++ user/get.go
@@ -1,9 +1,9 @@
-package users
+package user
 
 import (
 	"encoding/json"
 	"errors"
-	"github.com/najemi-software/tidal-dl/v2/requests"
+	"github.com/najemi-software/tidal-dl/v3/requests"
 	"strconv"
 )