+++ user/types.go
@@ -1,0 +1,7 @@
+package user
+
+type User struct {
+	Id        int     `json:"id"`
+	FirstName *string `json:"firstName"`
+	LastName  *string `json:"lastName"`
+}