+++ CHANGELOG.md
@@ -5,9 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [3.0.0] - 2025-02-20
+
+### Changed
+
+- `users` packaged renamed to `user`
+- `user` first and last name types set to `*string`
+
 ## [2.1.1] - 2025-02-06
 
-- Fix error-causing bug in InteractiveLogin
+### Fixed
+
+- Error-causing bug in InteractiveLogin
 
 ## [2.1.0] - 2025-02-03